+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2026-04-18
 - `16:00` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
+- `16:00` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `15:57` -- **Added** `apps/desktopProbe/src/server/pushover.ts`
 - `15:57` -- **Updated** `apps/desktopProbe/src/lib/types.ts`